Added statistics feature. Added success and failure notifications for authentication and game events.
Get games returns all the games played by the user. Looping through the
returned array of games it calculates how many games 'X' won, 'O' won
and the number of draws. It also counts how many games are still in
progress.

Using a status bar at the bottom of the display, let's the user know
the success or failure of authentication and game events.